task 12 correction

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -10,9 +10,9 @@
  */
 void print_number(int n)
 {
-	long p;
+	int p;
 	int i;
-	long num;
+	int num;
 
 	num = n;
